small fixes

+++ pkg/vault/test/foundry/RecoveryMode.t.sol
@@ -114,7 +114,7 @@ contract RecoveryModeTest is BaseVaultTest {
         require(vault.isVaultPaused(), "Vault should be paused initially");
         require(vault.isPoolPaused(pool) == false, "Pool should not be paused initially");
 
-        uint256 bufferPeriodEndTime = vault.getBufferPeriodEndTime();
+        uint32 bufferPeriodEndTime = vault.getBufferPeriodEndTime();
 
         // Ensure we are in the permissionless period of the Vault.
         skip(bufferPeriodEndTime);
@@ -148,7 +148,7 @@ contract RecoveryModeTest is BaseVaultTest {
         assertFalse(vault.isPoolInRecoveryMode(pool), "Pool should not be in Recovery Mode after pausing");
 
         // Ensure we are in the permissionless period of the Pool.
-        (, , uint256 bufferPeriodEndTime, ) = vault.getPoolPausedState(pool);
+        (, , uint32 bufferPeriodEndTime, ) = vault.getPoolPausedState(pool);
 
         vm.warp(bufferPeriodEndTime + 1);
